Split out `url`
