+++ client/lib/media/utils/index.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import path from 'path';
-import photon from 'photon';
 import { includes, omitBy, startsWith, get } from 'lodash';
 import { isUri } from 'valid-url';
 import debugFactory from 'debug';
@@ -12,7 +11,6 @@ const debug = debugFactory( 'calypso:media' );
  * Internal dependencies
  */
 import { reduxGetState } from 'lib/redux-bridge';
-import resize from 'lib/resize-image-url';
 import {
 	MimeTypes,
 	VideoPressFileTypes,
@@ -36,61 +34,7 @@ const { uniqueId } = impureLodash;
  */
 const REGEXP_VIDEOPRESS_GUID = /^[a-z\d]+$/i;
 
-/**
- * Given a media object, returns a URL string to that media. Accepts
- * optional options to specify photon usage or a maximum image width.
- *
- * @param  {object} media   Media object
- * @param  {object} options Optional options, accepting a `photon` boolean,
- *                          `maxWidth` pixel value, `resize` string, or `size`.
- * @returns {string}         URL to the media
- */
-export function url( media, options ) {
-	if ( ! media ) {
-		return;
-	}
-
-	if ( media.transient ) {
-		return media.URL;
-	}
-
-	// We've found that some media can be corrupt with an unusable URL.
-	// Return early so attempts to parse the URL don't result in an error.
-	if ( ! media.URL ) {
-		return;
-	}
-
-	options = options || {};
-
-	if ( options.photon ) {
-		if ( options.maxWidth ) {
-			return photon( media.URL, { width: options.maxWidth } );
-		}
-		if ( options.resize ) {
-			return photon( media.URL, { resize: options.resize } );
-		}
-
-		return photon( media.URL );
-	}
-
-	if ( media.thumbnails && options.size in media.thumbnails ) {
-		return media.thumbnails[ options.size ];
-	}
-
-	if ( options.maxWidth ) {
-		return resize( media.URL, {
-			w: options.maxWidth,
-		} );
-	}
-
-	if ( options.resize ) {
-		return resize( media.URL, {
-			resize: options.resize,
-		} );
-	}
-
-	return media.URL;
-}
+export { url } from 'lib/media/utils/url';
 
 /**
  * Given a media string, File, or object, returns the file extension.
